refactor: 모달 관련 이벤트 수행함수 분리
